organize user settings components into folder
