+++ client/src/components/User/U_Settings/UserSettingsModal.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import { makeStyles } from "@material-ui/core/styles";
 import { Modal, Fade, Backdrop } from "@material-ui/core";
